Add type checking for meal data and extract form to a seperate file
