+++ src/types.tsx
@@ -1,0 +1,6 @@
+export type MealType = {
+  id: number;
+  food: string;
+  quantity: number;
+  calories: string;
+};
